refactor: divide controller dto with service dto
