+++ src/main/java/com/example/spinlog/article/repository/ArticleRepositoryCustom.java
@@ -1,7 +1,7 @@
 package com.example.spinlog.article.repository;
 
 import com.example.spinlog.article.service.request.SearchCond;
-import com.example.spinlog.article.dto.ViewArticleSumDto;
+import com.example.spinlog.article.service.response.ViewArticleSumDto;
 import com.example.spinlog.user.entity.User;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
