+++ src/main/java/com/example/spinlog/article/repository/ArticleRepositoryImpl.java
@@ -1,7 +1,7 @@
 package com.example.spinlog.article.repository;
 
 import com.example.spinlog.article.service.request.SearchCond;
-import com.example.spinlog.article.dto.ViewArticleSumDto;
+import com.example.spinlog.article.service.response.ViewArticleSumDto;
 import com.example.spinlog.article.entity.Emotion;
 import com.example.spinlog.article.entity.RegisterType;
 import com.example.spinlog.user.entity.User;
