+++ src/main/java/com/example/spinlog/article/service/response/ViewListResponseDto.java
@@ -1,20 +1,22 @@
 package com.example.spinlog.article.service.response;
 
 
-import com.example.spinlog.article.dto.ViewArticleSumDto;
 import com.example.spinlog.utils.NullDataConverter;
+import lombok.AccessLevel;
 import lombok.Builder;
-import lombok.Data;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import java.util.List;
 
-@Data
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class ViewListResponseDto {
     private boolean nextPage;
     private List<ViewArticleSumDto> spendList;
 
     @Builder
-    public ViewListResponseDto(boolean nextPage, List<ViewArticleSumDto> spendList) {
+    private ViewListResponseDto(boolean nextPage, List<ViewArticleSumDto> spendList) {
         this.nextPage = nextPage;
         this.spendList = NullDataConverter.convertList(spendList);
     }
