+++ src/test/java/com/example/spinlog/article/repository/ArticleRepositoryTest.java
@@ -2,7 +2,7 @@ package com.example.spinlog.article.repository;
 
 import com.example.spinlog.article.service.request.ArticleUpdateRequest;
 import com.example.spinlog.article.service.request.SearchCond;
-import com.example.spinlog.article.dto.ViewArticleSumDto;
+import com.example.spinlog.article.service.response.ViewArticleSumDto;
 import com.example.spinlog.article.entity.Article;
 import com.example.spinlog.user.entity.User;
 import com.example.spinlog.user.repository.UserRepository;
